Decoupled the classes and helpers used in run_info so that only the things that need to be changed are in the run_info.py file for ease of quick changes
